diffy-bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "diffy-bridge"
version = "0.1.0"
description = "Transparent access to .dfy files as standard video frames"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
authors = [{ name = "Diffy" }]
keywords = ["video", "compression", "codec", "diffy", "dfy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Video :: Conversion",
]

dependencies = [
    "numpy>=1.24",
    "Pillow>=9.0",
    "diffycodec",   # the core codec library
]

[project.optional-dependencies]
video = ["imageio[ffmpeg]>=2.28"]
torch = ["torch>=2.0"]
all = ["imageio[ffmpeg]>=2.28", "torch>=2.0"]

[project.scripts]
diffy = "diffy_bridge.cli:main"

[project.urls]
Homepage = "https://github.com/ibrue/Diffy"
Repository = "https://github.com/ibrue/Diffy"

[tool.setuptools.packages.find]
where = ["."]
include = ["diffy_bridge*"]
